Move UT to nativeSql folder
